[ci] Publish the package to pub.dev on push tags event (#32)
Although I checked out the tag to publish to pub.dev, it still failed
with the error:
```
The calling GitHub Action is not allowed to publish, because: publishing is only allowed from 'tag' refType, this token has 'branch' refType. See https://dart.dev/go/publishing-from-github
```

https://github.com/littleGnAl/glance/actions/runs/11040051016/job/30667088690

So I decided to publish the package to pub.dev on push tags event by
following the [official
doc](https://dart.dev/tools/pub/automated-publishing#configuring-a-github-action-workflow-for-publishing-to-pub-dev)
and test it in the next release.